PullRequest: 529 fix/dev-4.3.2-3217
Merge branch 'fix/dev-4.3.2-3216 of git@code.alipay.com:oceanbase/oceanbase-developer-center.git into dev-4.3.2

https://code.alipay.com/oceanbase/oceanbase-developer-center/pull_requests/529




* Fixes oceanbase/odc#3217
* Fixes oceanbase/odc#3346
* Fixes oceanbase/odc#3349
* Fixes oceanbase/odc#3217
* Fixes oceanbase/odc#3350
* Fixes oceanbase/odc#3359
* Fixes oceanbase/odc#3370
* Fixes oceanbase/odc#3363

+++ src/page/Project/Database/components/LogicDatabase/ManageLogicDatabase.tsx
@@ -18,7 +18,17 @@ import { ReactComponent as NewOpenSvg } from '@/svgr/newopen.svg';
 import { isLogicalDatabase } from '@/util/database';
 import { gotoSQLWorkspace } from '@/util/route';
 import Icon, { ExclamationCircleFilled } from '@ant-design/icons';
-import { Button, ConfigProvider, Descriptions, Drawer, Empty, message, Space, Tooltip } from 'antd';
+import {
+  Button,
+  ConfigProvider,
+  Descriptions,
+  Drawer,
+  Empty,
+  message,
+  Space,
+  Tooltip,
+  Modal,
+} from 'antd';
 import React, { useEffect, useRef, useState } from 'react';
 
 const getColumns = ({ logicalDatabaseId, reload }) => {
@@ -88,13 +98,17 @@ const getColumns = ({ logicalDatabaseId, reload }) => {
             <Action.Link
               key={'delete'}
               onClick={async () => {
-                const successful = await deleteLogicalTable(logicalDatabaseId, record.id);
-                if (successful) {
-                  message.success('删除成功');
-                  reload?.();
-                  return;
-                }
-                message.error('删除失败');
+                Modal.confirm({
+                  title: `确认要移除逻辑表${record?.name}?`,
+                  onOk: async () => {
+                    const successful = await deleteLogicalTable(logicalDatabaseId, record.id);
+                    if (successful) {
+                      message.success('移除成功');
+                      reload?.();
+                      return;
+                    }
+                  },
+                });
               }}
               disabled={record.physicalTableCount === 0}
               tooltip={record.physicalTableCount === 0 ? '存在物理表，暂不可移除' : ''}
